+++ doc/CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.0.2] - 2017-10-1
+- Fix dependency version
+
 ## [1.0.0] - 2017-10-1
 - Add gregorian calendar type
 - Add locale